Changed the score field to BigIntegerField from IntegerField, allowing a bigger number for scores when submissions time out

+++ django_project/judgy/models.py
@@ -79,7 +79,7 @@ class Submission(models.Model):
     problem = models.ForeignKey(Problem, on_delete=models.CASCADE)
     team = models.ForeignKey(Team, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    score = models.IntegerField()
+    score = models.BigIntegerField()
     time = models.DateTimeField(auto_now_add=True)
 
 class Notification(models.Model):
